requirements.txt
ahttp==1.1.1
aiohttp==3.8.1
Django==3.2.15
oracledb==1.3.0
PyMySQL==1.0.2
requests==2.24.0
xlrd==1.2.0
jsonpath
django_apscheduler
apscheduler
django_redis
python-hpilo
redfish
pywinrm
Pillow
lxml
lxml-html-clean
uwsgi
